Ported to Windows with Qt 6.3.1
Removed WiringPi header
Added Portaudio and FFTW3 libraries and headers
Commented/removed code related to EMLID GPS.
Modified .pro for win32.
Changed QMYSQL to QSQLITE.

+++ src/mainwindow.cpp
@@ -433,7 +433,7 @@ void MainWindow::warningAccept() {
     char *confName = nameBA.data();
     char datetimeString[40];
     std::strftime(datetimeString, sizeof(datetimeString), "%Y-%m-%d %H:%M:%S", std::gmtime(&currentDatetime));
-    DataLogger::Configuration conf = {confName, datetimeString, sampleSize, sampleFrequency};
+    DataLogger::Configuration conf = {confName, datetimeString, (unsigned int)sampleSize, sampleFrequency};
     emit logConfiguration(conf);
     stateInstance->newLog();
     ui->statusLabel->setText(stateInstance->stateString[stateInstance->getState()]);
